Refactor JavaScript modules and remove unused code
- Moved `onDocumentReady` and `schedule` functions from `lib/init.js` to `app.js`.
- Added `getCssHsl` function to `metrics-charts.js` and removed it from `lib/dom.js`.
- Deleted `lib/dom.js` and `lib/init.js` as they contained unused functions.
- Updated `json-tree.js` to include a new implementation of `decodeHtmlEntities`.
- Refactored JSON parsing logic in `stream-guard.js` to use `parseJson` function.
